requirements.txt
PyJWT

# aiohttp stack
aiohttp
cchardet
aiodns
aiohttp_debugtoolbar

asynctest
pyyaml

# For cache
werkzeug


# Management
django
